// interfaces/form.interface.ts
export interface FormData {
    company?: string
    salutation: string
    firstName: string
    lastName: string
    address: string
    postalCode: string
    city: string
    country: string
    email: string
    phone?: string
    width?: string
    depth?: string
    height?: string
    roofCovering?: string
    message?: string
    privacyPolicyAccepted: boolean
}